add coerce function, move package
