+++ hail/src/main/scala/is/hail/expr/types/TypeWithRequiredness.scala
@@ -1,8 +1,8 @@
-package is.hail.expr.types.virtual
+package is.hail.expr.types
 
 import is.hail.annotations.Annotation
-import is.hail.expr.types.TableType
 import is.hail.expr.types.physical._
+import is.hail.expr.types.virtual._
 import is.hail.utils.{FastSeq, Interval}
 import org.apache.spark.sql.Row
